Install Gatsby
Adds the Gatsby static website generator for React.js. This allows the
usage of React.js without running Node.js on the server.

This is based off of the Gatsby Material-UI Starter template. It handles
installing and setting up the Material-UI library to work with Gatsby.

Additionally, adds the NOTICE.md file to provide attribution to the
template as required by its MIT license.

Addresses #2

Signed-off-by: Grant Iraci <grantira@buffalo.edu>