Change to Create React App
Change the framework / build process from Gatsby to Create React App.
This allows easy client-side rendering support, which Gatsby doesn't do
easily. It still allows for the site to be compiled down to a set of
statically hosted files. The main difference is that this doesn't do the
React.js rendering ahead of time.

Signed-off-by: Grant Iraci <grantira@buffalo.edu>