Start Linking between Machine Pages and Select Page.
This commit is the start of the Machine View page. Adds links
from machine cards on select page to Machine View page (dfa_home.js),
as well as back button from Machine View page to select page.

This commit also adds pages and components directories to src to better
organize and modularize code.

Dependency of @material_ui/icons/ was added, which is reflected in
modifications to package.json and package-lock.json.

Addresses #4

Signed-off-by: Matt Eichhorn <maeichho@buffalo.edu>