Change document title on web pages.
This commit changes the document title of the home page to
"ADAM: Another Drawer of Alan's Machines". The title of
each machine select page is the name of the machine itself.

Previously the title of the App Bar was changed,
this commit additionally changes the title of the webpage.

Addresses #4

TODO: When machines have multiple pages, the titles of these
pages should be different. The proposed format for names is:
    View: <Machine Name>
    Edit: <Machine Name>
    Run: <Machine Name>

Signed-off-by: Blake Hurlburt <blakehur@buffalo.edu>