Fix errors in App class related to state change
State needs to be set synchronously in the constructor
of the App class, otherwise it is null the first time
the page is rendered. Member functions in the App
class need to be created as bindings to anonymous
functions in order to call methods belonging to
the super class.

Addresses #4

Signed-off-by: Blake Hurlburt <blakehur@buffalo.edu>