Update dialog to allow submission of valid names
Before submission, machine names need to be validated.
A new machine name may not be empty, and may not have
the same name as an existing machine. Currently the
submission dialog closes when a valid name is
submitted, but does not add a new card to the
list of machines.

Addresses #16

Signed-off-by: Matthew Eichhorn <maeichho@buffalo.edu>
Signed-off-by: Blake Hurlburt <blakehur@buffalo.edu>